Fixed sorting in Stats > Growth > Top sources
ref https://linear.app/ghost/issue/PROD-2126/
- rebuilt stats endpoint as `top-sources-growth` instead of top referrers to be a little more clear
- refactored source mapping to be done on back end which is necessary for the limit to be applied

We should refactor these changes to put the source mapping into its own package. We already have @tryghost/referrerParser, and maybe that would be a decent home, or else we should create something similar. It doesn't make sense to maintain this in multiple locations.
