+++ apps/posts/src/hooks/usePostReferrers.ts
@@ -1,119 +1,7 @@
 import moment from 'moment';
-import {getSymbol} from '@tryghost/admin-x-framework';
+import {getSymbol, normalizeSource} from '@tryghost/admin-x-framework';
 import {useMemo} from 'react';
 import {useMrrHistory, usePostGrowthStats as usePostGrowthStatsAPI, usePostReferrers as usePostReferrersAPI} from '@tryghost/admin-x-framework/api/stats';
-
-// Source normalization mapping - matches Tinybird's mv_hits.pipe logic
-// Using Map for O(1) lookup performance
-const SOURCE_MAPPING = new Map<string, string>([
-    // Social Media Consolidation
-    ['facebook', 'Facebook'],
-    ['www.facebook.com', 'Facebook'],
-    ['l.facebook.com', 'Facebook'],
-    ['lm.facebook.com', 'Facebook'],
-    ['m.facebook.com', 'Facebook'],
-    ['twitter', 'Twitter'],
-    ['x.com', 'Twitter'],
-    ['com.twitter.android', 'Twitter'],
-    ['go.bsky.app', 'Bluesky'],
-    ['bsky', 'Bluesky'],
-    ['bsky.app', 'Bluesky'],
-    ['linkedin', 'LinkedIn'],
-    ['www.linkedin.com', 'LinkedIn'],
-    ['linkedin.com', 'LinkedIn'],
-    ['instagram', 'Instagram'],
-    ['www.instagram.com', 'Instagram'],
-    ['instagram.com', 'Instagram'],
-    ['youtube', 'YouTube'],
-    ['www.youtube.com', 'YouTube'],
-    ['youtube.com', 'YouTube'],
-    ['m.youtube.com', 'YouTube'],
-    ['threads', 'Threads'],
-    ['www.threads.net', 'Threads'],
-    ['threads.net', 'Threads'],
-    ['tiktok', 'TikTok'],
-    ['www.tiktok.com', 'TikTok'],
-    ['tiktok.com', 'TikTok'],
-    ['pinterest', 'Pinterest'],
-    ['www.pinterest.com', 'Pinterest'],
-    ['pinterest.com', 'Pinterest'],
-    ['reddit', 'Reddit'],
-    ['www.reddit.com', 'Reddit'],
-    ['reddit.com', 'Reddit'],
-    ['whatsapp', 'WhatsApp'],
-    ['whatsapp.com', 'WhatsApp'],
-    ['www.whatsapp.com', 'WhatsApp'],
-    ['telegram', 'Telegram'],
-    ['telegram.org', 'Telegram'],
-    ['www.telegram.org', 'Telegram'],
-    ['t.me', 'Telegram'],
-    ['news.ycombinator.com', 'Hacker News'],
-    ['substack', 'Substack'],
-    ['substack.com', 'Substack'],
-    ['www.substack.com', 'Substack'],
-    ['medium', 'Medium'],
-    ['medium.com', 'Medium'],
-    ['www.medium.com', 'Medium'],
-    
-    // Search Engines
-    ['google', 'Google'],
-    ['www.google.com', 'Google'],
-    ['google.com', 'Google'],
-    ['bing', 'Bing'],
-    ['www.bing.com', 'Bing'],
-    ['bing.com', 'Bing'],
-    ['yahoo', 'Yahoo'],
-    ['www.yahoo.com', 'Yahoo'],
-    ['yahoo.com', 'Yahoo'],
-    ['search.yahoo.com', 'Yahoo'],
-    ['duckduckgo', 'DuckDuckGo'],
-    ['duckduckgo.com', 'DuckDuckGo'],
-    ['www.duckduckgo.com', 'DuckDuckGo'],
-    ['search.brave.com', 'Brave Search'],
-    ['yandex', 'Yandex'],
-    ['yandex.com', 'Yandex'],
-    ['www.yandex.com', 'Yandex'],
-    ['baidu', 'Baidu'],
-    ['baidu.com', 'Baidu'],
-    ['www.baidu.com', 'Baidu'],
-    ['ecosia', 'Ecosia'],
-    ['www.ecosia.org', 'Ecosia'],
-    ['ecosia.org', 'Ecosia'],
-    
-    // Email Platforms
-    ['gmail', 'Gmail'],
-    ['mail.google.com', 'Gmail'],
-    ['gmail.com', 'Gmail'],
-    ['outlook', 'Outlook'],
-    ['outlook.live.com', 'Outlook'],
-    ['outlook.com', 'Outlook'],
-    ['hotmail.com', 'Outlook'],
-    ['mail.yahoo.com', 'Yahoo Mail'],
-    ['ymail.com', 'Yahoo Mail'],
-    ['icloud.com', 'Apple Mail'],
-    ['me.com', 'Apple Mail'],
-    ['mac.com', 'Apple Mail'],
-    
-    // News Aggregators
-    ['news.google.com', 'Google News'],
-    ['apple.news', 'Apple News'],
-    ['flipboard', 'Flipboard'],
-    ['flipboard.com', 'Flipboard'],
-    ['www.flipboard.com', 'Flipboard'],
-    ['smartnews', 'SmartNews'],
-    ['smartnews.com', 'SmartNews'],
-    ['www.smartnews.com', 'SmartNews']
-]);
-
-const normalizeSource = (source: string): string => {
-    if (!source || source === '') {
-        return '';
-    }
-    
-    // Case-insensitive lookup
-    const lowerSource = source.toLowerCase();
-    return SOURCE_MAPPING.get(lowerSource) || source;
-};
 
 // Helper function to convert range to date parameters
 export const getRangeDates = (rangeInDays: number) => {
