+++ ghost/core/core/server/api/endpoints/stats.js
@@ -110,36 +110,6 @@ const controller = {
             return await statsService.api.getReferrersHistory();
         }
     },
-    referrersHistoryWithRange: {
-        headers: {
-            cacheInvalidate: false
-        },
-        options: [
-            'date_from',
-            'date_to',
-            'timezone',
-            'member_status'
-        ],
-        permissions: {
-            docName: 'posts',
-            method: 'browse'
-        },
-        cache: statsService.cache,
-        generateCacheKeyData(frame) {
-            return {
-                method: 'referrersHistoryWithRange',
-                options: {
-                    date_from: frame.options.date_from,
-                    date_to: frame.options.date_to,
-                    timezone: frame.options.timezone,
-                    member_status: frame.options.member_status
-                }
-            };
-        },
-        async query(frame) {
-            return await statsService.api.getReferrersHistoryWithRange(frame.options.date_from, frame.options.date_to);
-        }
-    },
     topContent: {
         headers: {
             cacheInvalidate: false
@@ -530,6 +500,45 @@ const controller = {
             return {
                 data: [memberCounts]
             };
+        }
+    },
+    topSourcesGrowth: {
+        headers: {
+            cacheInvalidate: false
+        },
+        options: [
+            'order',
+            'limit', 
+            'date_from',
+            'date_to',
+            'timezone',
+            'member_status'
+        ],
+        permissions: {
+            docName: 'posts',
+            method: 'browse'
+        },
+        cache: statsService.cache,
+        generateCacheKeyData(frame) {
+            return {
+                method: 'topSourcesGrowth',
+                options: {
+                    order: frame.options.order,
+                    limit: frame.options.limit,
+                    date_from: frame.options.date_from,
+                    date_to: frame.options.date_to,
+                    timezone: frame.options.timezone,
+                    member_status: frame.options.member_status
+                }
+            };
+        },
+        async query(frame) {
+            return await statsService.api.getTopSourcesWithRange(
+                frame.options.date_from, 
+                frame.options.date_to, 
+                frame.options.order || 'free_members desc',
+                frame.options.limit || 50
+            );
         }
     }
 };
