+++ ghost/core/core/server/services/stats/ReferrersStatsService.js
@@ -1,5 +1,121 @@
 const moment = require('moment');
 
+// Source normalization mapping - consolidated from frontend apps
+const SOURCE_NORMALIZATION_MAP = new Map([
+    // Social Media Consolidation
+    ['facebook', 'Facebook'],
+    ['www.facebook.com', 'Facebook'],
+    ['l.facebook.com', 'Facebook'],
+    ['lm.facebook.com', 'Facebook'],
+    ['m.facebook.com', 'Facebook'],
+    ['twitter', 'Twitter'],
+    ['x.com', 'Twitter'],
+    ['com.twitter.android', 'Twitter'],
+    ['go.bsky.app', 'Bluesky'],
+    ['bsky', 'Bluesky'],
+    ['bsky.app', 'Bluesky'],
+    ['linkedin', 'LinkedIn'],
+    ['www.linkedin.com', 'LinkedIn'],
+    ['linkedin.com', 'LinkedIn'],
+    ['instagram', 'Instagram'],
+    ['www.instagram.com', 'Instagram'],
+    ['instagram.com', 'Instagram'],
+    ['youtube', 'YouTube'],
+    ['www.youtube.com', 'YouTube'],
+    ['youtube.com', 'YouTube'],
+    ['m.youtube.com', 'YouTube'],
+    ['threads', 'Threads'],
+    ['www.threads.net', 'Threads'],
+    ['threads.net', 'Threads'],
+    ['tiktok', 'TikTok'],
+    ['www.tiktok.com', 'TikTok'],
+    ['tiktok.com', 'TikTok'],
+    ['pinterest', 'Pinterest'],
+    ['www.pinterest.com', 'Pinterest'],
+    ['pinterest.com', 'Pinterest'],
+    ['reddit', 'Reddit'],
+    ['www.reddit.com', 'Reddit'],
+    ['reddit.com', 'Reddit'],
+    ['whatsapp', 'WhatsApp'],
+    ['whatsapp.com', 'WhatsApp'],
+    ['www.whatsapp.com', 'WhatsApp'],
+    ['telegram', 'Telegram'],
+    ['telegram.org', 'Telegram'],
+    ['www.telegram.org', 'Telegram'],
+    ['t.me', 'Telegram'],
+    ['news.ycombinator.com', 'Hacker News'],
+    ['substack', 'Substack'],
+    ['substack.com', 'Substack'],
+    ['www.substack.com', 'Substack'],
+    ['medium', 'Medium'],
+    ['medium.com', 'Medium'],
+    ['www.medium.com', 'Medium'],
+
+    // Search Engines
+    ['google', 'Google'],
+    ['www.google.com', 'Google'],
+    ['google.com', 'Google'],
+    ['bing', 'Bing'],
+    ['www.bing.com', 'Bing'],
+    ['bing.com', 'Bing'],
+    ['yahoo', 'Yahoo'],
+    ['www.yahoo.com', 'Yahoo'],
+    ['yahoo.com', 'Yahoo'],
+    ['search.yahoo.com', 'Yahoo'],
+    ['duckduckgo', 'DuckDuckGo'],
+    ['duckduckgo.com', 'DuckDuckGo'],
+    ['www.duckduckgo.com', 'DuckDuckGo'],
+    ['search.brave.com', 'Brave Search'],
+    ['yandex', 'Yandex'],
+    ['yandex.com', 'Yandex'],
+    ['www.yandex.com', 'Yandex'],
+    ['baidu', 'Baidu'],
+    ['baidu.com', 'Baidu'],
+    ['www.baidu.com', 'Baidu'],
+    ['ecosia', 'Ecosia'],
+    ['www.ecosia.org', 'Ecosia'],
+    ['ecosia.org', 'Ecosia'],
+
+    // Email Platforms
+    ['gmail', 'Gmail'],
+    ['mail.google.com', 'Gmail'],
+    ['gmail.com', 'Gmail'],
+    ['outlook', 'Outlook'],
+    ['outlook.live.com', 'Outlook'],
+    ['outlook.com', 'Outlook'],
+    ['hotmail.com', 'Outlook'],
+    ['mail.yahoo.com', 'Yahoo Mail'],
+    ['ymail.com', 'Yahoo Mail'],
+    ['icloud.com', 'Apple Mail'],
+    ['me.com', 'Apple Mail'],
+    ['mac.com', 'Apple Mail'],
+
+    // News Aggregators
+    ['news.google.com', 'Google News'],
+    ['apple.news', 'Apple News'],
+    ['flipboard', 'Flipboard'],
+    ['flipboard.com', 'Flipboard'],
+    ['www.flipboard.com', 'Flipboard'],
+    ['smartnews', 'SmartNews'],
+    ['smartnews.com', 'SmartNews'],
+    ['www.smartnews.com', 'SmartNews']
+]);
+
+/**
+ * Normalize source names to consistent display names
+ * @param {string|null} source - Raw source string from referrer data
+ * @returns {string} Normalized source name or 'Direct' for empty/null sources
+ */
+function normalizeSource(source) {
+    if (!source || source === '') {
+        return 'Direct';
+    }
+
+    // Case-insensitive lookup
+    const lowerSource = source.toLowerCase();
+    return SOURCE_NORMALIZATION_MAP.get(lowerSource) || source;
+}
+
 class ReferrersStatsService {
     /**
      * @param {object} deps
@@ -80,69 +196,6 @@ class ReferrersStatsService {
                 allEntries.push({
                     ...entry,
                     signups: 0,
-                    date: entryDate
-                });
-            }
-        });
-
-        // sort allEntries in date ascending format
-        allEntries.sort((a, b) => {
-            return moment(a.date).diff(moment(b.date));
-        });
-
-        return {
-            data: allEntries,
-            meta: {}
-        };
-    }
-
-    /**
-     * Return a list of all the attribution sources with date range filtering, including MRR data
-     * @param {string} startDate - Start date in YYYY-MM-DD format
-     * @param {string} endDate - End date in YYYY-MM-DD format
-     * @returns {Promise<{data: AttributionCountStatWithMrr[], meta: {}}>}
-     */
-    async getReferrersHistoryWithRange(startDate, endDate) {
-        const paidConversionEntries = await this.fetchPaidConversionSourcesWithRange(startDate, endDate);
-        const signupEntries = await this.fetchSignupSourcesWithRange(startDate, endDate);
-        const mrrEntries = await this.fetchMrrSourcesWithRange(startDate, endDate);
-
-        const allEntries = signupEntries.map((entry) => {
-            return {
-                ...entry,
-                paid_conversions: 0,
-                mrr: 0,
-                date: moment(entry.date).format('YYYY-MM-DD')
-            };
-        });
-
-        paidConversionEntries.forEach((entry) => {
-            const entryDate = moment(entry.date).format('YYYY-MM-DD');
-            const existingEntry = allEntries.find(e => e.source === entry.source && e.date === entryDate);
-
-            if (existingEntry) {
-                existingEntry.paid_conversions = entry.paid_conversions;
-            } else {
-                allEntries.push({
-                    ...entry,
-                    signups: 0,
-                    mrr: 0,
-                    date: entryDate
-                });
-            }
-        });
-
-        mrrEntries.forEach((entry) => {
-            const entryDate = moment(entry.date).format('YYYY-MM-DD');
-            const existingEntry = allEntries.find(e => e.source === entry.source && e.date === entryDate);
-
-            if (existingEntry) {
-                existingEntry.mrr = entry.mrr;
-            } else {
-                allEntries.push({
-                    ...entry,
-                    signups: 0,
-                    paid_conversions: 0,
                     date: entryDate
                 });
             }
@@ -267,6 +320,92 @@ class ReferrersStatsService {
             .orderBy('date');
 
         return rows;
+    }
+
+    /**
+     * Return aggregated attribution sources for a date range, grouped by source only (not by date)
+     * This is used for "Top Sources" tables that need server-side sorting
+     * @param {string} startDate - Start date in YYYY-MM-DD format
+     * @param {string} endDate - End date in YYYY-MM-DD format
+     * @param {string} [orderBy='signups desc'] - Sort order: 'signups desc', 'paid_conversions desc', 'mrr desc', 'source desc'
+     * @param {number} [limit=50] - Maximum number of sources to return
+     * @returns {Promise<{data: AttributionCountStatWithMrr[], meta: {}}>}
+     */
+    async getTopSourcesWithRange(startDate, endDate, orderBy = 'signups desc', limit = 50) {
+        const paidConversionEntries = await this.fetchPaidConversionSourcesWithRange(startDate, endDate);
+        const signupEntries = await this.fetchSignupSourcesWithRange(startDate, endDate);
+        const mrrEntries = await this.fetchMrrSourcesWithRange(startDate, endDate);
+
+        // Aggregate by source (not by date + source)
+        const sourceMap = new Map();
+
+        // Add signup data
+        signupEntries.forEach((entry) => {
+            const source = normalizeSource(entry.source);
+            const existing = sourceMap.get(source) || {source, signups: 0, paid_conversions: 0, mrr: 0};
+            existing.signups += entry.signups;
+            sourceMap.set(source, existing);
+        });
+
+        // Add paid conversion data
+        paidConversionEntries.forEach((entry) => {
+            const source = normalizeSource(entry.source);
+            const existing = sourceMap.get(source) || {source, signups: 0, paid_conversions: 0, mrr: 0};
+            existing.paid_conversions += entry.paid_conversions;
+            sourceMap.set(source, existing);
+        });
+
+        // Add MRR data
+        mrrEntries.forEach((entry) => {
+            const source = normalizeSource(entry.source);
+            const existing = sourceMap.get(source) || {source, signups: 0, paid_conversions: 0, mrr: 0};
+            existing.mrr += entry.mrr;
+            sourceMap.set(source, existing);
+        });
+
+        // Convert to array and sort
+        let results = Array.from(sourceMap.values());
+
+        // Apply sorting - only allow descending sorts for sources
+        const [field] = orderBy.split(' ');
+        
+        results.sort((a, b) => {
+            let valueA; let valueB;
+            
+            switch (field) {
+            case 'signups':
+                valueA = a.signups;
+                valueB = b.signups;
+                break;
+            case 'paid_conversions':
+                valueA = a.paid_conversions;
+                valueB = b.paid_conversions;
+                break;
+            case 'mrr':
+                valueA = a.mrr;
+                valueB = b.mrr;
+                break;
+            case 'source':
+                valueA = a.source.toLowerCase();
+                valueB = b.source.toLowerCase();
+                break;
+            default:
+                return 0;
+            }
+
+            // Always sort in descending order (highest to lowest)
+            return valueA < valueB ? 1 : valueA > valueB ? -1 : 0;
+        });
+
+        // Apply limit
+        if (limit && limit > 0) {
+            results = results.slice(0, limit);
+        }
+
+        return {
+            data: results,
+            meta: {}
+        };
     }
 }
