+++ ghost/core/core/server/web/api/endpoints/admin/routes.js
@@ -155,7 +155,6 @@ module.exports = function apiRoutes() {
     router.get('/stats/subscriptions', mw.authAdminApi, http(api.stats.subscriptions));
     router.get('/stats/referrers/posts/:id', mw.authAdminApi, http(api.stats.postReferrers));
     router.get('/stats/referrers', mw.authAdminApi, http(api.stats.referrersHistory));
-    router.get('/stats/referrers-growth', mw.authAdminApi, http(api.stats.referrersHistoryWithRange));
     if (labs.isSet('trafficAnalytics')) {
         router.get('/stats/posts/:id/stats', mw.authAdminApi, http(api.stats.postStats));
         router.get('/stats/top-posts', mw.authAdminApi, http(api.stats.topPosts));
@@ -167,6 +166,7 @@ module.exports = function apiRoutes() {
         router.get('/stats/subscriber-count', mw.authAdminApi, http(api.stats.subscriberCount));
         router.get('/stats/posts/:id/top-referrers', mw.authAdminApi, http(api.stats.postReferrersAlpha));
         router.get('/stats/posts/:id/growth', mw.authAdminApi, http(api.stats.postGrowthStats));
+        router.get('/stats/top-sources-growth', mw.authAdminApi, http(api.stats.topSourcesGrowth));
         router.post('/stats/posts-visitor-counts', mw.authAdminApi, http(api.stats.postsVisitorCounts));
         router.post('/stats/posts-member-counts', mw.authAdminApi, http(api.stats.postsMemberCounts));
     }
